refactor: cast id to Long (controller)

+++ src/main/kotlin/com/rinha-backend/KotlinRinhaBackend/controller/PersonController.kt
@@ -1,16 +1,9 @@
 package com.rinha.backend.KotlinRinhaBackend.controller
 
 import com.rinha.backend.KotlinRinhaBackend.model.PersonModel
-import org.springframework.web.bind.annotation.GetMapping
-import org.springframework.web.bind.annotation.PathVariable
-import org.springframework.web.bind.annotation.PostMapping
-import org.springframework.web.bind.annotation.RequestMapping
-import org.springframework.web.bind.annotation.RequestParam
-import org.springframework.web.bind.annotation.RestController
 import com.rinha.backend.KotlinRinhaBackend.service.PersonService
 import org.springframework.http.HttpStatus
-import org.springframework.web.bind.annotation.RequestBody
-import org.springframework.web.bind.annotation.ResponseStatus
+import org.springframework.web.bind.annotation.*
 
 @RestController
 @RequestMapping("")
@@ -24,11 +17,11 @@ class PersonController(private val service: PersonService) {
 
     @GetMapping("pessoas/{id}")
     fun searchPersonById(@PathVariable id: String): PersonModel {
-        return service.getPersonById(id)
+        return service.getPersonById(id.toLong())
     }
 
     @GetMapping("pessoas/")
-    fun searchPersonByFilter(@RequestParam("t") filter: String?): Collection<PersonModel>  {
+    fun searchPersonByFilter(@RequestParam("t") filter: String?): Collection<PersonModel> {
         if (filter == null) {
             return listOf()
         }
